requirements.txt
fastapi==0.115.9
pydantic==2.11.4
numpy==2.3.0
Pillow==11.1.0
requests==2.32.3
chromadb==1.0.12
sentence-transformers==4.1.0
transformers==4.52.4
torch==2.7.1
uvicorn==0.34.3